+++ example/src/lib/App.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import placeholder from '../img/placeholder.svg';
+import placeholder from '~img/placeholder.svg';
 
 export default function App(props) {
   const {children} = props;