+++ test/fixture/build-sass-loader-imports-ok/src/index.js
@@ -1,1 +1,1 @@
-require('./index.css');
+require('./index.scss');