+++ test/fixture/build-shortened-imports-ok/src/other.js
@@ -1,0 +1,1 @@
+module.exports = 'This is the shortened module!';